requirements.txt
asgiref==3.9.1
bcrypt==4.3.0
Django==5.2.4
djangorestframework==3.16.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
